==> notebooks/notebook_01_02_master.Rmd <==
---
title: "Week One, Class Two"
author: "ML"
date: "1/11/2022"
output: html_document
---

## Getting Started

To begin, copy all the text on this screen. Then open RStudio, and in the "File" menu select "New File" then "R Markdown". Add a title like "SOCI 1230, Week One, Class Two", include your name in the "Author" box, select "PDF" as the Default Output Format, and click Ok. A default markdown file will open. Delete all the text in the default file below the header (everything after line 7) and paste in the text you copied to your clipboard. Make sure that "R Markdown" is selected in the file type drop down menu in the bottom right corner of this pane.

## Load Packages

```{r load tidyverse}
library(tidyverse)
```

## Morning Recap

Reminder that the ggplot2 cheatsheet is on our Canvas page. Go to Pages>R Resources and you'll see several other cheatsheets as well.

This morning we used the diamonds dataset to make boxplots of price by clarity:

```{r boxplot recap}
ggplot(diamonds, aes(x = clarity, y = price)) +
  geom_boxplot()
```

The values of price are in dollars, but the y-axis labels are not. Use the `scales` package to bring in more options for axis labeling.

```{r scales package}
#install.packages("scales") # put a hashtag in front of this line after installing
library(scales)
```

```{r change axis label format}
ggplot(diamonds, aes(x = clarity, y = price)) +
  geom_boxplot() +
  scale_y_continuous(labels = dollar_format()) # to change axis labels to dollars
```

## Load Our Data

We'll continue using Table 2 from Opportunity Insights' Mobility Report Cards paper. Load the data as a data frame called `mobility`.

```{r load data and packages}
mobility <-  read.csv("https://opportunityinsights.org/wp-content/uploads/2018/04/mrc_table2.csv", stringsAsFactors = TRUE)
```

Can you make a boxplot showing the distribution of `par_q1` by `tier_name`?

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r parq1 boxplot example}
ggplot(mobility, aes(x = tier_name, y = par_q1)) +
  geom_boxplot() 
```

How could we improve this?

The `label_wrap()` function from the scales package breaks up long axis labels.

```{r flip axes}
ggplot(mobility, aes(x = tier_name, y = par_q1)) +
  geom_boxplot() +
  scale_x_discrete(labels = label_wrap(10)) # For max 10 characters (in multiple words) on each label line
```

The `reorder()` function sorts the axis labels by the value of another variable.

```{r order axis by value of another variable}
ggplot(mobility, aes(x = reorder(tier_name, par_q1), # order tier_name by par_q1
                     y = par_q1)) +
  geom_boxplot() +
  scale_x_discrete(labels = label_wrap(10)) +
  labs(x = "",
       y = "Propotion of students from bottom 20%",
       title = "Proportion of students from bottom 20% by college tier",
       caption = "Data come from Opportunity Insights")
```

## Reviewing Mobility Tables

We ended class yesterday by making a function to create a mobility table for any college in our data frame. Let's re-run the function so we can use it again.

```{r function setup}
mobility_table <- function(x) {
  df <- mobility |> 
    filter(name == x)
  
  pq1 <- c(df$kq1_cond_parq1,
           df$kq2_cond_parq1,
           df$kq3_cond_parq1,
           df$kq4_cond_parq1,
           df$kq5_cond_parq1)
  
  pq2 <- c(df$kq1_cond_parq2,
           df$kq2_cond_parq2,
           df$kq3_cond_parq2,
           df$kq4_cond_parq2,
           df$kq5_cond_parq2)
  
  pq3 <- c(df$kq1_cond_parq3,
           df$kq2_cond_parq3,
           df$kq3_cond_parq3,
           df$kq4_cond_parq3,
           df$kq5_cond_parq3)
  
  pq4 <- c(df$kq1_cond_parq4,
           df$kq2_cond_parq4,
           df$kq3_cond_parq4,
           df$kq4_cond_parq4,
           df$kq5_cond_parq4)
  
  pq5 <- c(df$kq1_cond_parq5,
           df$kq2_cond_parq5,
           df$kq3_cond_parq5,
           df$kq4_cond_parq5,
           df$kq5_cond_parq5)
  
  college_table <- rbind(pq1, pq2, pq3, pq4, pq5)
  
  colnames(college_table) <- c("kq1", "kq2", "kq3", "kq4", "kq5")
  
  round(college_table,3)
}
```

And use this function to get Middlebury's mobility table.

```{r run the mobility table function}
mobility_table("City College Of New York - CUNY")
```

Remember that the top right cell is what Chetty et al call the *success rate*: the proportion of students from the bottom quintile of the parent's income distribution who end up in the top quintile of the kid's income distribution.

What other colleges did you try? Any colleges from the Leonhardt column?

## Adapting the Mobility Table Function

In the mobility table above, each row sums to 1. That helps us summarize the destination quintiles for students from each origin quintile. But that does not help us know what proportion of *all* students at a college are in each cell. We need to be able to compare these other proportions to know which colleges have the highest proportions of their students experiencing mobility. 

How can we adapt the function we already created to write a new function (called `mobility_table_2`) that will give us a table showing the proportion of all students at a college in each pq-kq cell? When you copy and paste the function code from above, note that the chunk's name needs to change for the file to knit properly.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r function setup 2}
mobility_table_2 <- function(x) { # change the name of the object
  df <- mobility |> 
    filter(name == x) # Filter!
  
  pq1 <- c(df$kq1_cond_parq1,
           df$kq2_cond_parq1,
           df$kq3_cond_parq1,
           df$kq4_cond_parq1,
           df$kq5_cond_parq1) * df$par_q1 # par_q1 is the *access rate*
  
  pq2 <- c(df$kq1_cond_parq2,
           df$kq2_cond_parq2,
           df$kq3_cond_parq2,
           df$kq4_cond_parq2,
           df$kq5_cond_parq2) * df$par_q2
  
  pq3 <- c(df$kq1_cond_parq3,
           df$kq2_cond_parq3,
           df$kq3_cond_parq3,
           df$kq4_cond_parq3,
           df$kq5_cond_parq3) * df$par_q3
  
  pq4 <- c(df$kq1_cond_parq4,
           df$kq2_cond_parq4,
           df$kq3_cond_parq4,
           df$kq4_cond_parq4,
           df$kq5_cond_parq4) * df$par_q4
  
  pq5 <- c(df$kq1_cond_parq5,
           df$kq2_cond_parq5,
           df$kq3_cond_parq5,
           df$kq4_cond_parq5,
           df$kq5_cond_parq5) * df$par_q5
  
  college_table <- rbind(pq1, pq2, pq3, pq4, pq5)
  
  colnames(college_table) <- c("kq1", "kq2", "kq3", "kq4", "kq5")
  
  round(college_table,3)
}
```

Run this function the same way as the original function by adding a college's name.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r run the weighted mobility function}
mobility_table_2("La Sierra University")
```

The top right cell of this table is what Chetty et al call the *mobility rate*: the proportion of all students at an institution from the bottom quintile of the parent's income distribution who end up in the top quintile of the kid's income distribution.

Does anyone know how to show that all the cell values add to 1?

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
sum(mobility_table_2("Middlebury College"))
```


Get this weighted mobility table for City College in the CUNY system.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
mobility_table_2("City College Of New York - CUNY")
```


## Comparing Mobility Rates

Which colleges have the highest mobility rates? There is already a variable in the data frame measuring the mobility rate: `mr_kq5_pq1`. The easiest thing to do at an early stage of analysis is to open the spreadsheet view of the data frame and sort by the values of that variable.

Tomorrow we'll see the `select()` function which helps us choose to keep certain columns. Here's a preview:

```{r introducing the select function}
high_mobility_rate <- select(mobility, # the full data frame
                             name, tier_name, mr_kq5_pq1) # the variables to keep
```

How can we use some of the geom types we saw this morning to visualize the relationship across the variables we have been using? Try a scatterplot of access rates (`par_q1`) by success rates (`kq5_cond_parq1`) with points shaded by their mobility rates (`mr_kq5_pq1`).

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###
```{r scatterplot setup}
ggplot(mobility, aes(x = par_q1, y = kq5_cond_parq1, color = mr_kq5_pq1)) +
  geom_point()
```

How could we improve this visualization? Let's try changing the scale of the colors, changing the theme, and adding labels.

```{r scatterplot scale and theme}
ggplot(mobility, aes(x = par_q1, y = kq5_cond_parq1, color = mr_kq5_pq1)) +
  geom_point() +
  scale_color_viridis_c(option = "inferno") + # a scale built in to ggplot
  theme_dark() + # helpful for the light viridis colors to stand out
  labs(x = "Proportion of students from bottom 20%",
       y = "Proportion of students from bottom 20% \nwho end up in top 20%", # \n for new line
       color = "Mobility Rate") # color = color legend
```

We might also want to highlight specific points based on values of a variable. I like the `gghighlight` package for this. Install and load the package.

```{r gghighlight }
#install.packages("gghighlight") # Put a hashtag in front after installing
library(gghighlight)
```

The `gghighlight()` function works like the `filter()` function. Simply assert the conditions for the highlighted layer.

```{r gghighlight example}
ggplot(mobility, aes(x = par_q1, y = kq5_cond_parq1)) + geom_point() +
  gghighlight(name == "Middlebury College")
```

Add the `label_key =` option for a label based on the value of a variable:

```{r gghighlight label example}
ggplot(mobility, aes(x = par_q1, y = kq5_cond_parq1)) + geom_point() +
  gghighlight(name == "Middlebury College", 
              label_key = name)
```

What other points might be interesting to highlight? Can we find the NESCAC schools?

```{r gghighlight NESCAC}
ggplot(mobility, aes(x = par_q1, y = kq5_cond_parq1)) + geom_point() +
  gghighlight(tier_name=="Other elite schools (public and private)" & 
                count<750 & #needs to be higher to include Tufts
                state %in% c("VT", "NH", "ME", "MA", "NY", "RI"), 
              label_key = name)
```


## "America's Great Working Class Colleges"

What do we need to replicate this table from Leonhardt's column: 

https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/notes_slides/1230_figures/mrc_upward_mobility_top10.png

Pay attention to the notes at the top and the bottom. What functions do they make you think of? What variables do we already have in our dataframe? What new variables do we need?

We'll put our table in a new data frame called `top10mobility` based on the existing `mobility` data frame. Start by using `filter()` to choose the rows we want to keep.

```{r filter example}
top10mobility <- filter(mobility,
                        count>500 & par_q1>.1)
```

Then there are a few new things to learn to finish the table...

```{r other functions to replicate table}
top10mobility |> # this "pipe" allows us to chain multiple functions together
  mutate(kq345_cond_parq1 = kq3_cond_parq1 + kq4_cond_parq1 + kq5_cond_parq1) |>  # to create a new variable
  arrange(desc(kq345_cond_parq1)) |> # to sort by highest value to lowest value of a variable
  select(name, kq345_cond_parq1) |> # to keep specific columns
  top_n(10, kq345_cond_parq1) # to keep a certain number of observations, sorted by a variable
```
