==> notebooks/notebook_03_01_master.Rmd <==
---
name: "Week Three, Class One"
author: "ML"
output: pdf_output
---

# Set up

```{r}
library(gganimate)
library(tidyverse)
library(scales)
library(ggthemes)
library(weights)
library(kableExtra)
library(patchwork)
```

# Animation Debrief

Any thoughts on how to use animate functions with our data?

```{r}
chetty_cohorts <- read_csv("https://opportunityinsights.org/wp-content/uploads/2018/04/mrc_table3.csv")

chetty <- read_csv("https://opportunityinsights.org/wp-content/uploads/2018/04/mrc_table2.csv")
```

```{r}
mrc_subset <- chetty_cohorts |> 
  select(name, cohort, count, tier_name, type, iclevel, par_q1, par_q2, par_q3)

animation1 <- mrc_subset |> 
  mutate(par_q123 = par_q1 + par_q2 + par_q3,
         iclevel_with_none = ifelse(is.na(iclevel), 4, iclevel),
         iclevel_with_none = factor(iclevel_with_none,
                                    levels = c(4,3,2,1),
                                    labels = c("None",
                                               "Less than two year",
                                              "Two year", 
                                              "Four year"))) |> 
  group_by(cohort, iclevel_with_none) |> 
  summarise(level_n = sum(count, na.rm = TRUE),
            parq123_mean = mean(par_q123, na.rm = TRUE),
            parq123_n = parq123_mean * level_n) |> 
  mutate(parq123_total = sum(parq123_n),
         prop_parq123 = parq123_n / parq123_total) |> 
  ggplot(aes(x = iclevel_with_none, 
             y = prop_parq123, 
             fill = factor(iclevel_with_none))) +
  geom_col() + guides(fill = "none") + theme_economist() +
  theme(axis.title.x = element_blank(),
        axis.text.x = element_text(size = 14)) +
  scale_x_discrete(labels = label_wrap(10)) +
  labs(title = "Postsecondary Destinations Of Bottom 60% \nOf Income Distribution",
       subtitle = 'Year: {frame_time}') +
  transition_time(cohort)
```

```{r}
animate(animation1,
        nframes = 12,
        fps = 1)
```


```{r}
nescac <- c("Amherst College", "Bates College", "Bowdoin College", "Colby College", "Connecticut College", "Hamilton College", "Middlebury College", "Tufts University", "Trinity College of Hartford, CT", "Wesleyan University", "Williams College")

animation2 <- mrc_subset |> 
  filter(name %in% nescac) |> 
  mutate(par_q123 = par_q1 + par_q2 + par_q3) |> 
  group_by(cohort, name) |> 
  summarise(parq123_mean = mean(par_q123, na.rm = TRUE)) |> 
  arrange(cohort, -parq123_mean) |> 
  mutate(rank = row_number()) |> 
  filter(rank <=3 | name == "Middlebury College") |> 
  ggplot(aes(rank,
             parq123_mean, group = name,
             fill = as.factor(name), 
             color = as.factor(name))) +
  geom_col() +
  geom_text(aes(y = 0, label = paste(name, " ")), vjust = 0.2, hjust = 1) +
  geom_text(aes(y=parq123_mean,label = as.character(round(parq123_mean,3)), hjust=-1)) +
  coord_flip(clip = "off", expand = FALSE) +
  scale_x_reverse() +
  guides(color = "none", fill = "none") + 
  theme(axis.line=element_blank(),
        axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks=element_blank(),
        axis.title.x=element_blank(),
         axis.title.y=element_blank(),
        panel.background=element_blank(),
        panel.border=element_blank(),
        panel.grid.major=element_blank(),
        panel.grid.minor=element_blank(),
        panel.grid.major.x = element_line( size=.1, color="grey" ),
        panel.grid.minor.x = element_line( size=.1, color="grey" ),
        plot.title=element_text(size=25, hjust=0.5, face="bold", colour="grey", vjust=-1),
        plot.subtitle=element_text(size=18, hjust=0.5, face="italic", color="grey"),
        plot.caption =element_text(size=8, hjust=0.5, face="italic", color="grey"),
        plot.background=element_blank(),
       plot.margin = margin(2,2, 2, 4, "cm")) 

animation2 +
  transition_states(cohort, 
                    transition_length = 0, 
                    state_length = 20) +
  view_follow(fixed_x = FALSE)  +
  labs(title = 'Year : {closest_state}',  
       subtitle  =  "Proportion of students from bottom quintile by year",
       caption  = "parq1 variable from Opportunity Insights") +
  #enter_fade() 
  #exit_shrink() +
  enter_appear() + exit_fly()
```


# Assignment Debrief

```{r}
css_means <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/css_means_mobility_withno.csv")

tfs_means <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/tfs_means_mobility_withno.csv")
```

## Part One - TFS

The `REASON10` question in the TFS captures responses to the question of whether the respondent attended college "to learn more about things that interest me." Find the WEIGHTED correlations between each response to this question and `k_rank` for each type of college. Summarize your findings in a table (formatted with kable) and interpret your findings in a few sentences.

### Start by setting up data.

```{r, warning = FALSE, error = FALSE, message = FALSE}
tfs_cor_df <- tfs_means |> 
  select(campus_id, 
         n_responses, 
         type,
         k_rank,
         starts_with("REASON10")) |> 
  mutate(n_responses_nona = n_responses * (1-REASON10.propna)) |> 
  pivot_longer(names_to = "response_level",
               values_to = "prop",
               (starts_with("REASON10") & !contains("propna"))) |> 
  mutate(response_level = str_remove(response_level, "_mean")) |> 
  mutate(response_level = str_remove(response_level, "REASON10.")) |> 
  mutate(response_level = factor(response_level,
                                 levels = c("notimportant",
                                            "somewhatimportant",
                                            "veryimportant"),
                                 labels = c("Not important",
                                            "Somewhat important",
                                            "Very important"))) 
```

### Then summarise

```{r, warning = FALSE, error = FALSE, message = FALSE}
tfs_cor_summary <- tfs_cor_df |>
  group_by(type, response_level) |>
  summarise(wtd_cor_krank = wtd.cor(prop,
                              k_rank,
                              w = n_responses_nona)[1]) |>
  mutate(across(where(is.numeric),round,3))
```

### Then kable
```{r}
tfs_cor_summary |> 
  pivot_wider(names_from = "response_level",
              values_from = "wtd_cor_krank") |> 
  rename("Institution Type" = type) |> 
  kbl(booktabs = TRUE,
       align = "lccc",
      caption = "Weighted Correlations: k_rank and REASON10") |> 
    add_header_above(c(" " = 1, "Importance" = 3)) |> 
  kable_paper() # Nice theme to see kable output in notebook
```

## Part One - CSS

The `HPW15` question in the CSS captures responses to the question of how often the respondent socialized with friends in person. Combine the responses into these categories: 2 hours or less, 3 hours - 10 hours, 11 hours - 20 hours, 21 hours or more. Find the correlations between each of the re-categorized responses to this question and `k_rank` for each type of college. Make one figure that includes the ***FOUR*** scatterplots. 

### Set up data (watch new categories)
```{r, warning = FALSE, error = FALSE, message = FALSE}
css_cor_df <- css_means |> 
  select(campus_id, 
         n_responses, 
         type,
         k_rank,
         starts_with("HPW15")) |> 
  mutate(n_responses_nona = n_responses * (1-HPW15.propna)) |> 
  mutate(HPW_0_2 = HPW15.hours0_mean + HPW15.hours0to1_mean + HPW15.hours1to2_mean,
         HPW_3_10 = HPW15.hours3to5_mean + HPW15.hours6to10_mean,
         HPW_11_20 = HPW15.hours11to15_mean + HPW15.hours16to20_mean,
         HPW_21_plus = HPW15.hours21plus_mean) |> 
  select(-starts_with("HPW15")) |> 
  pivot_longer(names_to = "response_level",
               values_to = "prop",
               (starts_with("HPW_"))) |> 
  mutate(response_level = factor(response_level,
                                 levels = c("HPW_0_2",
                                            "HPW_3_10",
                                            "HPW_11_20",
                                            "HPW_21_plus"),
                                 labels = c("2 Hours or Less",
                                            "3 - 10 Hours",
                                            "11 - 20 Hours",
                                            "21 Hours or More"))) 
```

### Set up summarise
```{r}
css_cor_summary <- css_cor_df |> 
  group_by(type, response_level) |>
  summarise(wtd_cor_krank = wtd.cor(prop,
                              k_rank,
                              w = n_responses_nona)[1]) |>
  mutate(across(where(is.numeric),round,3))
```

### Set up plot
```{r, warning = FALSE, error = FALSE, message = FALSE, fig.cap = "Figure Title"}
css_cor_df |> 
  group_by(response_level) |>
  ggplot(aes(x = prop, y = k_rank, 
             size = n_responses_nona,
             color = type)) +
  geom_point() + geom_smooth(method = "lm") +
  theme(legend.position = "bottom") + guides(size = "none") +
  facet_wrap(~response_level) 
```


## Part Two

Take all the correlations with `k_rank` that you found in Part One. Make one figure that combines plots of the coefficients (by type of college) in a plot similar to the Opportunity Insights summary plot.

### Make plots

```{r, echo = FALSE, warning = FALSE, error = FALSE, message = FALSE}
tfs_plot <- tfs_cor_summary |> 
  ggplot(aes(x = response_level,
             y = wtd_cor_krank,
             color = type,
             size = 3)) +
  geom_point(size = 3) + ylim(-1, 1) +
  coord_flip() + geom_hline(yintercept = 0, linetype = "dashed") +
  theme_tufte() +
    theme(legend.position = "bottom",
        title = element_text(size = 8)) +
  guides(size = "none") + # to get rid of size legend
  labs(color = "New Title")  # to change title of color legend

css_plot <- css_cor_summary |> 
  ggplot(aes(x = response_level,
             y = wtd_cor_krank,
             color = type)) +
  geom_point(size = 8) + ylim(-1, 1) +
  coord_flip() + geom_hline(yintercept = 0, linetype = "dashed") +
  theme_tufte() +
    theme(legend.position = "bottom",
        title = element_text(size = 8))
```

```{r}
tfs_plot / css_plot + 
  plot_annotation(title = "Correlations With Average Income Rank In Early Adulthood") +
  plot_layout(guides = "collect") & # For common legend
  theme(legend.position = "bottom") # Position common legend
```
An example with faceting...

```{r}
tfs_css_correlations <- bind_rows(tfs_cor_summary, 
                                  css_cor_summary) |> 
  mutate(question = ifelse(str_detect(response_level, "Hours"),
                           "CSS:\nHours Spent Socializing",
                           "TFS:\nImportance of Learning")) |> 
  ggplot(aes(x = response_level,
             y = wtd_cor_krank,
             color = type)) +
  geom_point(size = 3) + coord_flip() + 
  facet_grid(question~., scales = "free_y") +
  theme(legend.position = "bottom")

tfs_css_correlations
```

Possible extensions for app?

## Bonus (if you make good progress on the above parts during Thursday's class and want more practice)

The matched CSS-TFS file links surveys across years by student. Take individuals' responses to the GOAL questions referring to the importance of being well off financially in both surveys. Make a figure showing how the distribution of responses changed over time. The values are: 1 = "Not important", 2 = "Somewhat important", 3 = "Very important", 4 = "Essential". Write a few sentences explaining why you made your visualization decisions.

```{r, include = FALSE, warning = FALSE, error = FALSE, message = FALSE}

matched_df <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/matched_goal07.csv")

goal_labels <- c("Not important",
                 "Somewhat important",
                 "Very important",
                 "Essential")

matched_df <- matched_df |> 
  mutate(GOAL_CSS = factor(GOAL_CSS,
                           labels = goal_labels),
         GOAL_TFS = factor(GOAL_TFS,
                           labels = goal_labels))

goal_table <- round(prop.table(table(
  matched_df$GOAL_TFS,
  matched_df$GOAL_CSS),1),3)

goal_table |> 
kbl(booktabs = TRUE,
    align = c("cccc"),
    caption = "Distribution of GOAL Responses From First Year To Senior Year") |> 
  kable_paper() |> 
  add_header_above(c(" " = 1, "Senior Year Importance" = 4))
```

For a crosswalk of matched variable names in the TFS and CSS, use this file:

https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/matched_variables.csv


# Summarizing Associations

Need a way to see the distribution of a survey question (we'll use `FACPRV08` from the CSS) at points of a continuous mobility variable (we'll use `k_rank`). What are our options?

```{r}
summary(css_means$k_rank)
```


Let's try finding quantiles of the mobility variable using the `quantile()` function.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###
```{r}
quantile(css_means$k_rank, probs = c(.25, .5, .75))
```

There's a `wtd.quantile()` function (from the weights package) that's better for our data.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
wtd.quantile(css_means$k_rank, 
             css_means$n_responses, 
             probs = c(.25, .5, .75))
```

Are these really the quantiles we want?

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###
```{r}
wtd.quantile(chetty$k_rank, 
             chetty$count, 
             probs = c(.25, .5, .75))
```

Our dataset is a pretty small part of that distribution. What could be the problem?

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###
```{r}
chetty |> 
  filter(iclevel==1) |> # only four year colleges
  summarise(wtd = wtd.quantile(k_rank, count, probs = c(.25, .5, .75)))
```

Let's save the quantiles. And let's break the quantiles into bottom half, next 25%, and top 25%. What do we need to do that?

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
mobility_variable_quantiles <- chetty |> 
  filter(iclevel==1) |> 
  summarise(wtd = wtd.quantile(k_rank, count, probs = c(.5, .75)))
```

Now let's create our categories. We'll still use the `FACPRV08` variable.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###
```{r}
# Includes example of nested ifelse...

quantile_df <- css_means |> 
  select(campus_id, n_responses, starts_with("FACPRV08"), k_rank) |> 
  rename_with(~paste("propna"), .cols = ends_with("propna")) |>  
  mutate(n_responses_nona = n_responses * (1-propna)) |> 
  mutate(qtle = ifelse(k_rank < mobility_variable_quantiles$wtd[1], 1,
                       ifelse(k_rank > mobility_variable_quantiles$wtd[2], 3, 
                              2))) # Fill this in
quantile_df |> 
  count(qtle) 
```

Let's finish setting up the data frame for use

```{r}
quantile_df <-  quantile_df |> 
  pivot_longer(names_to = "response_level",
               values_to = "proportion",
               starts_with("FACPRV08")) |> 
  mutate(response_level = str_remove(response_level, "_mean"),
         response_level = str_extract(response_level, "[^.]*$")) |> 
  mutate(response_level = factor(response_level,
                                 levels = c("notatall",
                                            "occasionally",
                                            "frequently"),
                                 labels = c("Not at all",
                                            "Occasionally",
                                            "Frequently")))

```

Now we can summarise by qtle and response_level

```{r}
quantile_sum <- quantile_df |>  
  group_by(qtle, response_level) |> # Group by qtle as well now
  summarise(wtd.mean = wtd.mean(proportion, 
                                weights = n_responses_nona, 
                                na.rm = TRUE))

# Pivot wider if we want a table...
quantile_sum |> 
  pivot_wider(names_from = response_level,
              values_from = wtd.mean)
```

And plot that summary

```{r}
quantile_sum |> 
  ggplot(aes(x = response_level, y = wtd.mean, fill = response_level)) +
  geom_col() +
  labs(title = "",
       fill = "") +
  theme_tufte() +
  theme(legend.position = "none") + facet_grid(qtle~.) +
  scale_x_discrete(labels = label_wrap(10)) +
  geom_text(aes(label = round(wtd.mean, 3)),
            position = position_dodge(width = .9),
            vjust = 0) +
  ylim(0, .6)
```

Preview of using inputs to populate workflow

```{r}
css_labels <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/css_question_summary_labels.csv")

css_variable_labels <- css_labels |>
  mutate(response_level = str_extract(Summary, "[^.]*$")) |> 
  filter(Name == "FACPRV08") # question input

variable_levels <- css_variable_labels$response_level
variable_labels <- css_variable_labels$Label
variable_description <- css_variable_labels$Description[1]
```

```{r}
css_means |> 
  select(campus_id, n_responses, starts_with("FACPRV08"), k_rank) |> 
  rename_with(~paste0("propna"), .cols = ends_with("propna")) |>  
  mutate(n_responses_nona = n_responses * (1-propna),
         qtle = ifelse(k_rank < mobility_variable_quantiles$wtd[1], 1,
                       ifelse(k_rank > mobility_variable_quantiles$wtd[2], 3, 
                              2))) |> 
  pivot_longer(names_to = "response_level",
               values_to = "proportion",
               starts_with("FACPRV08")) |> 
  mutate(response_level = str_remove(response_level, "_mean"),
         response_level = str_extract(response_level, "[^.]*$")) |> 
  mutate(response_level = factor(response_level,
                                 levels = variable_levels, # Stored values
                                 labels = variable_labels)) |> # Stored values
  group_by(qtle, response_level) |>
  summarise(wtd.mean = wtd.mean(proportion, 
                                weights = n_responses_nona, 
                                na.rm = TRUE)) |> 
  ggplot(aes(x = response_level, y = wtd.mean, fill = response_level)) +
  geom_col() +
  labs(title = variable_description, # Stored value
       fill = "") +
  theme_tufte() +
  theme(legend.position = "none") + facet_grid(qtle~.) +
  scale_x_discrete(labels = label_wrap(10)) +
  geom_text(aes(label = round(wtd.mean, 3)),
            position = position_dodge(width = .9),
            vjust = 0) +
  ylim(0, .6)
```
