==> notebooks/notebook_03_03_master.Rmd <==
---
title: "Week Three, Class Three"
author: "ML"
date: "1/26/2022"
output: pdf_document
---

# Load the usual packages
```{r}
library(tidyverse)
library(scales)
library(ggthemes)
library(weights)
library(kableExtra)
library(patchwork)
library(DT)
```

# Load the data

## These files have the weighted correlations between every survey question (by response level) and every mobility variable (pooled and by institution type, identified in the `corr_type` variable).

```{r}
tfs_correlations <-
  read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/tfs_correlations_all.csv")

css_correlations <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/css_correlations_all.csv")
```

## These files have more detailed labels by question and response level

```{r}
tfs_labels <-
  read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/tfs_question_summary_labels.csv")

css_labels <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/css_question_summary_labels.csv")
```

Noticed something tricky about the `tfs_labels` and `css_labels` files. They only have pooled weighted proportions so we'll run into issues joining with the mega correlations files. Drop the `wtdprop` column so the private and public correlations will get labels when we join later.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
tfs_labels <- tfs_labels |> 
  select(-wtdprop)

css_labels <- css_labels |> 
  select(-wtdprop)
```

Now join each correlations file to its matching labels file. Call the new dfs `css_join` and `tfs_join`.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
css_join <- left_join(css_correlations, css_labels)
tfs_join <- left_join(tfs_correlations, tfs_labels)
```

# Summary Plot (almost there?)

We'll set up an example using the CSS correlations.

As a refresher, start by finding the 10 strongest positive correlations with k_rank. Use the pooled correlations.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
css_join |> 
  filter(corr_type == "Pooled") |> 
  slice_max(k_rank, n = 10) |>  
  select(Group, Description, Summary, k_rank, Label) |> 
  arrange(-k_rank)
```

And to plot...

```{r}
css_join |> 
  filter(corr_type == "Pooled") |> 
  slice_max(k_rank, n = 10) |> 
  mutate(corr_sign = "Positive") |> 
  select(Group, Description, Summary, k_rank, Label) |> 
  arrange(-k_rank) |> 
  ggplot(aes(x = reorder(Summary, k_rank),
             y = k_rank)) +
  geom_point() + coord_flip()
```

Yesterday we were trying to figure out if there's a better way to capture the topics in which these questions fall. What if we find the strongest positive correlations within each `Group`? Call the dataframe `css_positive`.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
css_positive <- css_join |> 
  filter(corr_type == "Pooled") |> 
  group_by(Group) |> # The missing link!
  slice_max(k_rank, n = 1) |> 
  mutate(corr_sign = "Positive") |> 
  select(corr_sign, Group, Description, Summary, k_rank, Label) |> 
  arrange(-k_rank) |>
  mutate(abs_k_rank = abs(k_rank)) # Just in case we want it later...
```

That looks good. To be extra careful, before finding the max by group let's drop the "Constructed Scales" group and any response level with "cant" (for can't rate which might lead to weird results later).

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
css_positive <- css_join |> 
  filter(corr_type == "Pooled", 
         Group!="Constructed Scales",
         !str_detect(Summary, "cant")) |> 
  group_by(Group) |> # The missing link!
  slice_max(k_rank, n = 1) |> 
  mutate(corr_sign = "Positive") |> 
  select(corr_sign, Group, Description, Summary, k_rank, Label) |> 
  arrange(-k_rank) |>
  mutate(abs_k_rank = abs(k_rank)) # Just in case we want it later...
```

Do the same thing for the strongest negative correlations within each `Group`. Call the df `css_negative`.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
css_negative <- css_join |> 
  filter(corr_type == "Pooled", 
         Group!="Constructed Scales",
         !str_detect(Summary, "cant")) |> 
  group_by(Group) |> # The missing link!
  slice_min(k_rank, n = 1) |>  
  mutate(corr_sign = "Negative") |> 
  select(corr_sign, Group, Description, Summary, k_rank, Label) |> 
  arrange(-k_rank) |>
  mutate(abs_k_rank = abs(k_rank)) # Just in case we want it later...
```

Now we can connect the positive and negative dataframes into one data frame called `css_correlations_sign`.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
css_correlations_sign <- bind_rows(css_positive,
                                   css_negative)
```

Try making a plot showing points for the strongest positive and negative correlations in each group.

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###

```{r}
css_correlations_sign |> 
  ggplot(aes(x = Group, y = k_rank, color = corr_sign)) +
  geom_point(size = 3) + 
  coord_flip() + theme(legend.position = "bottom")
```

Any ways to improve this?

### REPLACE THIS LINE WITH YOUR CODE CHUNK ###
```{r}
css_correlations_sign |> 
  ggplot(aes(x = Group, y = k_rank, color = corr_sign)) +
  geom_point(size = 3) + 
  coord_flip() + theme(legend.position = "bottom") +
  scale_color_manual(values = c("red", "forest green")) # manual colors
```


By default, the groups are ordered alphabetically. We can force the ordering by pulling the order of the groups when sorted by positive correlation and then use that order as our levels.

```{r}
css_group_levels <- css_correlations_sign |> 
  filter(corr_sign=="Positive") |> 
  arrange(k_rank) |> 
  pull(Group) # we saw this in today's morning session

css_group_levels
```

```{r}
css_plot <- css_correlations_sign |> 
  mutate(Group = factor(Group,
                        levels = css_group_levels)) |> 
  ggplot(aes(x = Group, y = k_rank, color = corr_sign)) +
  geom_point(size = 3) + 
  coord_flip() + theme(legend.position = "bottom") +
  scale_color_manual(values = c("red", "forest green"))

css_plot
```

Still would like some way to identify questions. What if we make the plot interactive with plotly? Let's see what we have so far.

```{r}
library(plotly)

ggplotly(css_plot)
```


Not a bad start but we can do better by customizing the tooltip text and moving the legend.

```{r}
css_plot <- css_correlations_sign |> 
  mutate(Group = factor(Group,
                        levels = css_group_levels)) |> 
  ggplot(aes(x = Group, y = k_rank, 
             color = corr_sign,
             text = paste("Group: ", Group, "\n", # text for custom tooltip
                          "Question: ", Description,"\n",
                          "Response: ", Label, "\n",
                          "Correlation: ", round(k_rank,3)))) +
  geom_point(size = 3) + 
  coord_flip() + theme(legend.position = "bottom") +
  scale_color_manual(values = c("red", "forest green"))

ggplotly(css_plot, tooltip = "text") |> 
  layout(legend = list(orientation = "h", x = .25, y = -0.2))
```

See here for setting up tooltip font formatting:

https://plotly-r.com/controlling-tooltips.html#styling

```{r}
font <- list(
  family = "Arial",
  size = 20,
  color = "white"
)
label <- list(
  #bgcolor = size,
  bordercolor = "transparent",
  font = font
)

```

```{r}

ggplotly(css_plot, tooltip = "text") |> 
  layout(legend = list(orientation = "h", x = .25, y = -0.2)) |> 
  style(hoverlabel = label) |> 
  layout(font = font)
```

Before tomorrow's afternoon session: do the same thing for the strongest negative correlations within each `Group`. Call the df `tfs_negative`.


# Collapsing the question groups

Google sheets are here:

- For TFS: https://docs.google.com/spreadsheets/d/1xaNCVmMfqfrk0MDWJEulUc0yLz35nOgbYgxuzRVPfuA/edit#gid=0

- For CSS: https://docs.google.com/spreadsheets/d/1y2ywEeHoe6iLyUVzLaH3Wtxl61rDdzq9xN9DpwQd08M/edit?pli=1#gid=0

- For Mobility: https://docs.google.com/spreadsheets/d/18YH5URpcThSj7QiEnVtLaXD1QDgHQBhIPVb-IzMDoWw/edit?usp=sharing

Edit the Google sheet directly. Make changes to the `Group_rev` column (for TFS and CSS) or the `Description`, `Keep`, and `Group` columns (for mobility) then we'll load those changes into R when you are done.

# Working with Google Sheets in R

```{r}
#install.packages("googlesheets4") 
library(googlesheets4)
gs4_deauth() # changes permissions so you don't need to authenticate to read public file
```

### TFS

```{r}
tfs_gsheet <- "https://docs.google.com/spreadsheets/d/1xaNCVmMfqfrk0MDWJEulUc0yLz35nOgbYgxuzRVPfuA/edit#gid=0"
tfs_groups <- read_sheet(tfs_gsheet)

tfs_labels <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/tfs_question_summary_labels.csv")

tfs_question_summary2 <- left_join(tfs_labels, tfs_groups)
```

### CSS

```{r}
css_gsheet <- "https://docs.google.com/spreadsheets/d/1y2ywEeHoe6iLyUVzLaH3Wtxl61rDdzq9xN9DpwQd08M/edit?pli=1#gid=0"
css_groups <- read_sheet(css_gsheet)

css_labels <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/css_question_summary_labels.csv")

css_question_summary2 <- left_join(css_labels, css_groups)
```

### Mobility

```{r}
mobility_gsheet <- "https://docs.google.com/spreadsheets/d/18YH5URpcThSj7QiEnVtLaXD1QDgHQBhIPVb-IzMDoWw/edit?usp=sharing"
mobility_groups <- read_sheet(mobility_gsheet)
```
