==> notebooks/notebook_04_01.Rmd <==
# Load the usual packages
```{r}
library(tidyverse)
library(scales)
library(ggthemes)
library(weights)
library(kableExtra)
library(patchwork)
library(DT)
library(plotly)
```

# Load the data

## Before final cleanup, here are the files our app uses:

```{r}
# These include revised groups
tfs_df <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/tfs_question_summary_labels_shiny.csv")

css_df <- read_csv("../data/css_question_summary_labels.csv")

# These include the constructed mobility variables:
# k_rank_cond_parq123 and kq345_cond_parq1

tfs_mobility <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/tfs_means_mobility_withno_shiny.csv")

css_mobility <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/css_means_mobility_withno_shiny.csv")

# These include revised groups
tfs_correlations_join <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/tfs_correlations_join_shiny.csv")

css_correlations_join <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/css_correlations_join_shiny.csv")

# No changes here
chetty_fouryr <- read_csv("https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/data/chetty_fouryr_shiny.csv")
```

Let's make some final changes before writing these files to the app folder

# TFS Descriptions

## Hanwen, Andy, and Sophia suggested these changes:

```{r}
tfs_df <- tfs_df |>
  mutate(Description = str_remove(Description, "Choose to Attend:"))|>
  mutate(Description = str_remove(Description, "Choice:"))|>
  mutate(Description = str_remove(Description, "Goal:"))|>
  mutate(Description = str_remove(Description, "Future Act:"))|>
  mutate(Description = str_remove(Description, "Reason Attend:"))|>
  mutate(Description = str_remove(Description, "Self Rating:"))|>
  mutate(Description = str_remove(Description, "View:")) |> 
  mutate(Description = str_remove(Description, "[:print:]+:\\s"))
```

## Anything else to fix?

### Removing the colons above worked well. Let's also remove them in the "Your sex:" description

```{r}
tfs_df <- tfs_df |> 
  mutate(Description = ifelse(str_detect(Description, "Your sex:"),
                              "Sex Composition", Description))
```

### Let's add the RACE_ETH variable we created in notebook_02_02.

```{r}
race_variables <- tfs_df |> 
  filter(str_detect(Name, regex("race", ignore_case = TRUE)), # to ignore case
         str_detect(Summary, "yes"))

race_variables <- race_variables |> 
  filter(!str_detect(Name, "GROUP"))

race_variables <- race_variables |> 
  mutate(Label = Description) |> 
  mutate(Name = "RACE_ETHNICITY")

tfs_df <- tfs_df |> 
  filter(!str_detect(Name, "RACE"))

tfs_df <- bind_rows(tfs_df, race_variables)

tfs_df <- tfs_df |> 
  mutate(Description = str_trim(Description, side = c("left")))

```

# CSS Descriptions

## Dilay, Evan, and Jaime: Any changes?

## What else to fix?

### Let's remove the "Family Background And Demographics" and "Scales" group:

```{r}
css_df <- css_df |> 
  filter(!str_detect(Group_rev, "Family|Scales"))
```

## Group Suggestions:

```{r}
css_summary2 <- css_df |> 
  mutate(Description = ifelse(Name == "ACT05", # [If this is true do str_replace]
                              str_replace(Description, # where to find the string
                                   "professors", # what string to find
                                   "professor's"), # how to replace match
                              Description))
```

```{r}
css_summary2 <- css_summary2 |> 
  mutate(find_colon = str_count(Description, ":"))
```

The default is to put new variables after the last column. Use `relocate()` to move columns. The default location for `relocate()` is the first column, but you can adjust that with the `.before =` or `.after =` options.

```{r}
css_summary2 <- css_summary2 |> 
  relocate(find_colon, .after = Description)
```

Check a table of the `find_colon` variable to see how bad the problem is.

```{r}
table(css_summary2$find_colon)
```


So we have 15 rows that require additional care. What are our options?


One possibility is to break the `Description` values into two pieces: a string before a colon and a string after (if there is a colon). Then we can use those two pieces to repair the existing values. We can do that with the `separate()` function:

```{r}
css_summary2 <- css_summary2 |> 
  separate(Description, # variable to separate
           c("before_colon", "after_colon"), # the two new columns to create
           sep = "[:]", # where to separate
           remove = FALSE) # don't remove the variable you are separating
```

Our operations above left some whitespace at the start of the strings in the `after_colon` variable. Trim it.

```{r}
css_summary2 <- css_summary2 |> 
  mutate(after_colon = str_trim(after_colon, side = c("left")))
```

Now try to change the existing Description values that have a colon to only take the text after the colon.

```{r}
css_summary2 <- css_summary2 |> 
  mutate(Description = ifelse(find_colon==1, # if this is true...
                              after_colon, # do this...
                              Description)) # if not true do this
```

```{r}
css_summary2 <- css_summary2 |> 
  mutate(Description = ifelse(Name == "ACT05", # [If this is true do str_replace]
                              str_replace(Description, # where to find the string
                                   "professors", # what string to find
                                   "professor's"), # how to replace match
                              Description)) # [If not true do this]
```


```{r}
css_summary2 <- css_summary2 |>
  mutate(before_colon = str_replace_all(before_colon, "Act in College", "Have you")) |>
  mutate(before_colon = str_replace_all(before_colon, "Act in Past Year", "In the past year, have you")) |>
  mutate(before_colon = str_replace_all(before_colon, "Act", "Have you")) |>
  mutate(before_colon = str_replace_all(before_colon, "CSV Performed", "Have you done community/volunteer service")) |>
  mutate(before_colon = str_replace_all(before_colon, "Satisfaction", "How satisfied are you with the")) |>
  mutate(before_colon = str_replace_all(before_colon, "Faculty Provide", "Has the faculty provided")) |>
  mutate(before_colon = str_replace_all(before_colon, "Change", "How have you changed your")) |>
  mutate(before_colon = str_replace_all(before_colon, "Self Rating", "Rate your")) |>
  mutate(before_colon = str_replace_all(before_colon, "View", "Do you agree/disagree that")) |>
  mutate(before_colon = str_replace_all(before_colon, "Success", "Are you")) |>
  mutate(before_colon = str_replace_all(before_colon, "Hours per Week", "How many hours per week did you spend")) |>
  mutate(before_colon = str_replace_all(before_colon, "Goal", "How important to you is")) |>
  mutate(before_colon = str_replace_all(before_colon, "Plans", "Do you plan on"))

```

```{r}
css_summary2 <- css_summary2 |>
  mutate(after_colon = str_to_lower(after_colon))
```

```{r}
css_summary2 <- css_summary2 |>
  mutate(Qmark = "?") |>
  mutate(Description2 = ifelse(find_colon == 0, 
                               before_colon, 
                               paste(before_colon, after_colon, sep = " "))) |>
  mutate(Description2 = ifelse(find_colon == 0, 
                               Description2, 
                               paste(Description2, Qmark, sep = ""))) 
```

```{r}
#css_summary2 <- css_summary2 |> 
  #mutate(Name = str_remove_all(Name, "_9406")) |>
  #mutate(Summary = str_remove_all(Summary, "_9406"))
```


```{r}
css_summary2 <- css_summary2 |> 
  select(Name, Description2, Summary, 
         Label_rank, wtdprop, Survey, Survey_Name,
         Label, Group, Group_rev) |> 
  rename(Description = Description2)
```

Remove question mark if rate is in the description.

```{r}
css_summary3 <- css_summary2 |> 
  mutate(Description = ifelse(str_detect(Description, "Rate"),
                              str_remove(Description, "[?]"), Description))
```



# Write all these changes to csv files in your app folder

```{r}
write.csv(tfs_df, "../data/tfs_question_summary_labels_shiny.csv", row.names = FALSE)
write.csv(css_summary3, "../data/css_question_summary_labels_shiny.csv", row.names = FALSE)

write.csv(tfs_mobility, "../data/tfs_means_mobility_withno_shiny.csv", row.names = FALSE)

write.csv(css_mobility, "../data/css_means_mobility_withno_shiny.csv", row.names = FALSE)

write.csv(tfs_correlations_join, "../data/tfs_correlations_join_shiny.csv", row.names = FALSE) 
write.csv(css_correlations_join, "../data/css_correlations_join_shiny.csv", row.names = FALSE) 

write.csv(chetty_fouryr, "../data/chetty_fouryr_shiny.csv", row.names = FALSE)
```

# The App!

Install a couple more packages for everything to run properly:

```{r}
install.packages(c("shinyfullscreen", "shinythemes"))
```


The most up to date version is here:

https://raw.githubusercontent.com/mjclawrence/soci1230_w22/main/final_project/app.R
